# be/docs/swagger.yaml
basePath: /
definitions:
...
  api.Credentials:
    properties:
      login:
        example: admin
        type: string
      pwd:
        example: password123
        type: string
    type: object
  api.ErrorResponse:
    properties:
      code:
        description: Error code for frontend i18n
        type: string
      message:
        description: Fallback message in English
        type: string
      params:
        additionalProperties:
          type: string
        description: Dynamic parameters for interpolation
        type: object
    type: object
...
host: localhost:1971
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is the API documentation for the ρBee (rhobee) backend server.
    It provides endpoints for user authentication, navigation, file management, and
    integration with Ollama AI services.
  license:
    name: Apache 2.0
    url: https://opensource.org/licenses/Apache-2.0
  termsOfService: http://swagger.io/terms/
  title: ρBee (rhobee) API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and receive JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/api.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: token and user info
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: User login
      tags:
      - auth
  /logout:
    post:
      description: Invalidate the current JWT token. The token must be provided in
        the Authorization header.
      produces:
      - application/json
      responses:
        "200":
          description: logout message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...